requirements.txt
alembic==1.12.0
bcrypt==4.0.1
bleach==6.1.0
blinker==1.6.3
click==8.1.7
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
webencodings==0.5.1
Werkzeug==2.3.3